backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy==2.0.36
python-multipart==0.0.20
websockets==14.1
httpx==0.28.1
anthropic==0.42.0
google-cloud-storage==2.19.0
google-cloud-firestore==2.21.0
google-api-python-client==2.159.0
google-auth==2.37.0
aiofiles==24.1.0
python-dotenv==1.0.1
jinja2==3.1.5
